+++ HISTORY.rst
@@ -16,6 +16,8 @@ XXXX-XX-XX
 **Bug fixes**
 
 - 715_: do not print exception on import time in case cpu_times() fails.
+- 1277_: [OSX] available and used memory (psutil.virtual_memory()) metrics are
+  not accurate.
 - 1294_: [Windows] psutil.Process().connections() may sometimes fail with
   intermittent 0xC0000001.  (patch by Sylvain Duchesne)
 - 1320_: [AIX] system CPU times (psutil.cpu_times()) were being reported with